--- a/crates/rolldown_testing/src/utils.rs
+++ b/crates/rolldown_testing/src/utils.rs
@@ -1,19 +1,6 @@
 use std::{borrow::Cow, sync::LazyLock};
 
 use regex::Regex;
-use rolldown_common::BundlerOptions;
-
-pub fn assert_bundled(options: BundlerOptions) {
-  let result = tokio::runtime::Builder::new_multi_thread()
-    .enable_all()
-    .build()
-    .expect("Failed building the Runtime")
-    .block_on(async move {
-      let mut bundler = rolldown::Bundler::new(options).expect("Failed to create bundler");
-      bundler.generate().await
-    });
-  assert!(result.is_ok(), "Failed to bundle.");
-}
 
 #[macro_export]
 /// `std::file!` alternative that returns an absolute path.