closes https://github.com/rolldown/rolldown/issues/3754

Benchmark in apps/10000:

### `1.0.0-beta.38`

```
❯ hyperfine --warmup 3 --runs 10 --command-name "rolldown-latest" "../../node_modules/.bin/rolldown build --config rolldown.config.mjs"
Benchmark 1: rolldown-latest
  Time (mean ± σ):     553.6 ms ±  21.4 ms    [User: 1107.1 ms, System: 2948.2 ms]
  Range (min … max):   518.7 ms … 579.5 ms    10 runs
```

### `current`

```
❯ hyperfine --warmup 3 --runs 10 --command-name "rolldown-latest" "../../node_modules/.bin/rolldown build --config rolldown.config.mjs"
Benchmark 1: rolldown-latest
  Time (mean ± σ):     458.1 ms ±   6.5 ms    [User: 1052.2 ms, System: 1092.4 ms]
  Range (min … max):   447.5 ms … 468.8 ms    10 runs
```

I don't know why `codspeed` crashes, I added a feature for `codspeed` so that under `codspeed` it doesn't use `spawn_blocking`.